fix: edits to aeneas' blog post on oauth2+OIDC
Nice article, very  long. Maybe @vinckr  can help with some editorial decisions on what to cut, or maybe it's fine as is.

Question: on line 51: "be aware that there are two types of
implementing OAuth2 / OpenID Connect and this article only covers the latter:"
I find the word "types" to be ambiguous here - does it mean there are only two general contexts in which one would implement these protocols? Could you clarify here?

Line 89: "that is better suited for the 80%**." Want to clarify what the 80% is or do readers know? 

Line 630: "While there is still a long way to go for Ory to really become a new breed and standard, we already have a solid foundation!" I deleted breed bc idk what it means but if you like it, keep it.